+++ examples/macro-tests/tests/fail/33_service_handle_rejects_generic_path.rs
@@ -1,0 +1,3 @@
+fn main() {
+    let _ = service_daemon::service_handle!(worker::<u8>);
+}